Extract Parse and its impl for location
