+++ runtime/dancebox/Cargo.toml
@@ -49,6 +49,7 @@ pallet-xcm-core-buyer-runtime-api = { workspace = true }
 tanssi-relay-encoder = { workspace = true }
 tanssi-runtime-common = { workspace = true }
 tp-xcm-core-buyer = { workspace = true }
+tp-xcm-commons = { workspace = true }
 
 # Moonkit
 async-backing-primitives = { workspace = true }
@@ -272,6 +273,7 @@ std = [
 	"tp-author-noting-inherent/std",
 	"tp-traits/std",
 	"tp-xcm-core-buyer/std",
+	"tp-xcm-commons/std",
 	"westend-runtime-constants/std",
 	"westend-runtime/std",
 	"xcm-primitives/std",
