+++ runtime/dancebox/src/xcm_config.rs
@@ -63,6 +63,7 @@ use {
     },
     staging_xcm_executor::{traits::JustTry, XcmExecutor},
     tp_traits::ParathreadParams,
+    tp_xcm_commons::Parse,
 };
 
 parameter_types! {
@@ -410,41 +411,6 @@ impl frame_support::traits::ContainsPair<Asset, Location> for NativeAssetReserve
             }
         }
         false
-    }
-}
-
-pub trait Parse {
-    /// Returns the "chain" location part. It could be parent, sibling
-    /// parachain, or child parachain.
-    fn chain_part(&self) -> Option<Location>;
-    /// Returns "non-chain" location part.
-    fn non_chain_part(&self) -> Option<Location>;
-}
-
-impl Parse for Location {
-    fn chain_part(&self) -> Option<Location> {
-        match (self.parents, self.first_interior()) {
-            // sibling parachain
-            (1, Some(Parachain(id))) => Some(Location::new(1, [Parachain(*id)])),
-            // parent
-            (1, _) => Some(Location::parent()),
-            // children parachain
-            (0, Some(Parachain(id))) => Some(Location::new(0, [Parachain(*id)])),
-            _ => None,
-        }
-    }
-
-    fn non_chain_part(&self) -> Option<Location> {
-        let mut junctions = self.interior().clone();
-        while matches!(junctions.first(), Some(Parachain(_))) {
-            let _ = junctions.take_first();
-        }
-
-        if junctions != Here {
-            Some(Location::new(0, junctions))
-        } else {
-            None
-        }
     }
 }
